refactor: move video to its own folder
